Plug in r2d2-redis (not tokio-ified)
Anything that uses redis takes a connection out of the redis_pool
and a thread out of the thread_pool and communicates its response
back to the Service using a Future.

Everything else can either return synchronously or fire off something
else that returns a promise, and box that up for returning.